perf: improve collector algorithm to reduce memory use
The main idea is to use a number to track the number of sub references
in the intermediate tree structure (instead of a boolean). With that
information, we can more efficiently allocate indices, and reuse an
existing index in some cases.

Now, the resulting refs array should contain exactly the number of
required references.

closes #12