fix: use correct index for subchilds in some cases
To compile blocks, we traverse the content once to build a tree, then
traverse that tree later to build the final optimized block code. And
this is fine, except that until this commit, we kind of assumed that the
children would be encountered in the same order, which is not true.

So, before this commit, it was possible to have a situation where we
would have the children located at a wrong position. The fix is simple
then: just use the child index value instead of pushing into a list.